requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.2
djangorestframework==3.14.0
numpy==1.24.2
Pillow==9.5.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
